backend/embedding-pipeline/pyproject.toml
[project]
name = "embedding-pipeline"
version = "0.1.0"
description = "Embedding pipeline for Physical AI Robotics Book"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.7.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
